[SR-152] update address controller

+++ smart-rent/src/main/java/com/smartrent/dto/response/ListingResponse.java
@@ -67,7 +67,7 @@ public class ListingResponse {
 
     String priceUnit;
 
-    Integer addressId;
+    Long addressId;
 
     Float area;
